Refactoring, since now the tests pass
